+++ sdk/src/api/embeddedWallet.ts
@@ -25,7 +25,7 @@ import { EvmProvider, Provider } from '../wallets/evm';
 import { announceProvider, openfortProviderInfo } from '../wallets/evm/provider/eip6963';
 import { TypedDataPayload } from '../wallets/evm/types';
 import { signMessage } from '../wallets/evm/walletHelpers';
-import { IframeManager } from '../wallets/iframeManager';
+import { IframeManager, SignerConfigureRequest } from '../wallets/iframeManager';
 import { ReactNativeMessenger } from '../wallets/messaging';
 import { WindowMessenger } from '../wallets/messaging/browserMessenger';
 import { MessagePoster } from '../wallets/types';
@@ -233,10 +233,15 @@ export class EmbeddedWalletApi {
 
     const auth = await Authentication.fromStorage(this.storage);
     const [signer, entropy] = await Promise.all([this.ensureSigner(), this.getEntropy(recoveryParams, auth!.player)]);
-    const account = await signer.configure({
+
+    const configureParams: SignerConfigureRequest = {
       chainId: params.chainId,
       entropy,
-    });
+      accountType: params.accountType ?? AccountTypeEnum.SMART_ACCOUNT,
+      chainType: params.chainType ?? ChainTypeEnum.EVM,
+    };
+
+    const account = await signer.configure(configureParams);
 
     return {
       id: account.id,