+++ sdk/src/types/types.ts
@@ -448,6 +448,8 @@ export type EmbeddedAccount = {
 export type EmbeddedAccountConfigureParams = {
   chainId?: number;
   recoveryParams: RecoveryParams;
+  chainType?: ChainTypeEnum;
+  accountType?: AccountTypeEnum;
 };
 
 export type EmbeddedAccountRecoverParams = {