+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.5] - 2025-08-29
+### Feat
+- Change API methods which configure method calls. Now it calls V2 methods
+- Dropped configuration attempts after "not configured" error from the iFrame
+
+## [0.10.4] - 2025-08-29
+### Feat
+- Export type ThirdPartyAuthConfiguration
+
+## [0.10.3] - 2025-08-29
+### Feat
+- Update third party auth getToken
+
+## [0.10.2] - 2025-08-29
+### Feat
+- change arguments for sign and export iFrame methods, make configuration mandatory
+
 ## [0.10.1] - 2025-08-26
 ### Feat
 - avoid redundant auth